+++ Core/GraphQL/API/sell_checkoutMyBasket.graphql.swift
@@ -25,11 +25,14 @@ public extension ApolloType {
       document.append("\n" + FragmentOrder.fragmentDefinition)
       document.append("\n" + FragmentProduct.fragmentDefinition)
       document.append("\n" + FragmentFulfilmentPoint.fragmentDefinition)
+      document.append("\n" + FragmentForm.fragmentDefinition)
       document.append("\n" + FragmentFulfilmentPointCategory.fragmentDefinition)
+      document.append("\n" + FragmentVenue.fragmentDefinition)
       document.append("\n" + FragmentTimeslot.fragmentDefinition)
       document.append("\n" + FragmentProductVariant.fragmentDefinition)
       document.append("\n" + FragmentProductModifierItemSelection.fragmentDefinition)
       document.append("\n" + FragmentProductModifierItem.fragmentDefinition)
+      document.append("\n" + FragmentUser.fragmentDefinition)
       return document
     }
