+++ Core/GraphQL/API/sell_fragment_fulfilmentPoint.graphql.swift
@@ -31,57 +31,7 @@ public extension ApolloType {
         }
         seatForm {
           __typename
-          id
-          reference
-          imageUrl
-          completionButtonUrl
-          completionButtonTitle
-          showCompletionDate
-          allowUpdate
-          refreshOnSuccess
-          validationIntegration
-          expiresIn
-          translations {
-            __typename
-            language
-            title
-            description
-            submitButtonTitle
-            skipButtonTitle
-            completionButtonTitle
-            completionTitle
-            completionDescription
-          }
-          fields {
-            __typename
-            type
-            key
-            validationRegex
-            required
-            sortId
-            autoFill {
-              __typename
-              type
-              field
-            }
-            translations {
-              __typename
-              title
-              language
-              description
-              completionDescription
-              skipButtonTitle
-              completionTitle
-              completionButtonTitle
-              submitButtonTitle
-            }
-            selectOptions {
-              __typename
-              title
-              value
-              iconUrl
-            }
-          }
+          ...FragmentForm
         }
         categories {
           __typename
@@ -89,25 +39,7 @@ public extension ApolloType {
         }
         venue {
           __typename
-          id
-          label
-          name
-          status
-          isDefault
-          description
-          imageUrl
-          mapImageUrl
-          geoLat
-          geoLong
-          geoLatNorthWest
-          geoLongNorthWest
-          geoLatSouthEast
-          geoLongSouthEast
-          city
-          externalId
-          createdAt
-          updatedAt
-          venueIconUrl
+          ...FragmentVenue
         }
         timeslots {
           __typename
@@ -368,18 +300,7 @@ public extension ApolloType {
       public static var selections: [GraphQLSelection] {
         return [
           GraphQLField("__typename", type: .nonNull(.scalar(String.self))),
-          GraphQLField("id", type: .nonNull(.scalar(GraphQLID.self))),
-          GraphQLField("reference", type: .scalar(String.self)),
-          GraphQLField("imageUrl", type: .scalar(String.self)),
-          GraphQLField("completionButtonUrl", type: .scalar(String.self)),
-          GraphQLField("completionButtonTitle", type: .scalar(String.self)),
-          GraphQLField("showCompletionDate", type: .scalar(Bool.self)),
-          GraphQLField("allowUpdate", type: .scalar(Bool.self)),
-          GraphQLField("refreshOnSuccess", type: .scalar(Bool.self)),
-          GraphQLField("validationIntegration", type: .scalar(String.self)),
-          GraphQLField("expiresIn", type: .scalar(Int.self)),
-          GraphQLField("translations", type: .nonNull(.list(.object(Translation.selections)))),
-          GraphQLField("fields", type: .list(.object(Field.selections))),
+          GraphQLFragmentSpread(FragmentForm.self),
         ]
       }
 
@@ -387,10 +308,6 @@ public extension ApolloType {
 
       public init(unsafeResultMap: ResultMap) {
         self.resultMap = unsafeResultMap
-      }
-
-      public init(id: GraphQLID, reference: String? = nil, imageUrl: String? = nil, completionButtonUrl: String? = nil, completionButtonTitle: String? = nil, showCompletionDate: Bool? = nil, allowUpdate: Bool? = nil, refreshOnSuccess: Bool? = nil, validationIntegration: String? = nil, expiresIn: Int? = nil, translations: [Translation?], fields: [Field?]? = nil) {
-        self.init(unsafeResultMap: ["__typename": "Form", "id": id, "reference": reference, "imageUrl": imageUrl, "completionButtonUrl": completionButtonUrl, "completionButtonTitle": completionButtonTitle, "showCompletionDate": showCompletionDate, "allowUpdate": allowUpdate, "refreshOnSuccess": refreshOnSuccess, "validationIntegration": validationIntegration, "expiresIn": expiresIn, "translations": translations.map { (value: Translation?) -> ResultMap? in value.flatMap { (value: Translation) -> ResultMap in value.resultMap } }, "fields": fields.flatMap { (value: [Field?]) -> [ResultMap?] in value.map { (value: Field?) -> ResultMap? in value.flatMap { (value: Field) -> ResultMap in value.resultMap } } }])
       }
 
       public var __typename: String {
@@ -402,545 +319,28 @@ public extension ApolloType {
         }
       }
 
-      public var id: GraphQLID {
+      public var fragments: Fragments {
         get {
-          return resultMap["id"]! as! GraphQLID
+          return Fragments(unsafeResultMap: resultMap)
         }
         set {
-          resultMap.updateValue(newValue, forKey: "id")
+          resultMap += newValue.resultMap
         }
       }
 
-      public var reference: String? {
-        get {
-          return resultMap["reference"] as? String
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "reference")
-        }
-      }
-
-      public var imageUrl: String? {
-        get {
-          return resultMap["imageUrl"] as? String
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "imageUrl")
-        }
-      }
-
-      public var completionButtonUrl: String? {
-        get {
-          return resultMap["completionButtonUrl"] as? String
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "completionButtonUrl")
-        }
-      }
-
-      public var completionButtonTitle: String? {
-        get {
-          return resultMap["completionButtonTitle"] as? String
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "completionButtonTitle")
-        }
-      }
-
-      public var showCompletionDate: Bool? {
-        get {
-          return resultMap["showCompletionDate"] as? Bool
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "showCompletionDate")
-        }
-      }
-
-      public var allowUpdate: Bool? {
-        get {
-          return resultMap["allowUpdate"] as? Bool
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "allowUpdate")
-        }
-      }
-
-      public var refreshOnSuccess: Bool? {
-        get {
-          return resultMap["refreshOnSuccess"] as? Bool
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "refreshOnSuccess")
-        }
-      }
-
-      public var validationIntegration: String? {
-        get {
-          return resultMap["validationIntegration"] as? String
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "validationIntegration")
-        }
-      }
-
-      public var expiresIn: Int? {
-        get {
-          return resultMap["expiresIn"] as? Int
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "expiresIn")
-        }
-      }
-
-      public var translations: [Translation?] {
-        get {
-          return (resultMap["translations"] as! [ResultMap?]).map { (value: ResultMap?) -> Translation? in value.flatMap { (value: ResultMap) -> Translation in Translation(unsafeResultMap: value) } }
-        }
-        set {
-          resultMap.updateValue(newValue.map { (value: Translation?) -> ResultMap? in value.flatMap { (value: Translation) -> ResultMap in value.resultMap } }, forKey: "translations")
-        }
-      }
-
-      public var fields: [Field?]? {
-        get {
-          return (resultMap["fields"] as? [ResultMap?]).flatMap { (value: [ResultMap?]) -> [Field?] in value.map { (value: ResultMap?) -> Field? in value.flatMap { (value: ResultMap) -> Field in Field(unsafeResultMap: value) } } }
-        }
-        set {
-          resultMap.updateValue(newValue.flatMap { (value: [Field?]) -> [ResultMap?] in value.map { (value: Field?) -> ResultMap? in value.flatMap { (value: Field) -> ResultMap in value.resultMap } } }, forKey: "fields")
-        }
-      }
-
-      public struct Translation: GraphQLSelectionSet {
-        public static let possibleTypes: [String] = ["FormTranslation"]
-
-        public static var selections: [GraphQLSelection] {
-          return [
-            GraphQLField("__typename", type: .nonNull(.scalar(String.self))),
-            GraphQLField("language", type: .nonNull(.scalar(String.self))),
-            GraphQLField("title", type: .scalar(String.self)),
-            GraphQLField("description", type: .scalar(String.self)),
-            GraphQLField("submitButtonTitle", type: .scalar(String.self)),
-            GraphQLField("skipButtonTitle", type: .scalar(String.self)),
-            GraphQLField("completionButtonTitle", type: .scalar(String.self)),
-            GraphQLField("completionTitle", type: .scalar(String.self)),
-            GraphQLField("completionDescription", type: .scalar(String.self)),
-          ]
-        }
-
+      public struct Fragments {
         public private(set) var resultMap: ResultMap
 
         public init(unsafeResultMap: ResultMap) {
           self.resultMap = unsafeResultMap
         }
 
-        public init(language: String, title: String? = nil, description: String? = nil, submitButtonTitle: String? = nil, skipButtonTitle: String? = nil, completionButtonTitle: String? = nil, completionTitle: String? = nil, completionDescription: String? = nil) {
-          self.init(unsafeResultMap: ["__typename": "FormTranslation", "language": language, "title": title, "description": description, "submitButtonTitle": submitButtonTitle, "skipButtonTitle": skipButtonTitle, "completionButtonTitle": completionButtonTitle, "completionTitle": completionTitle, "completionDescription": completionDescription])
-        }
-
-        public var __typename: String {
+        public var fragmentForm: FragmentForm {
           get {
-            return resultMap["__typename"]! as! String
+            return FragmentForm(unsafeResultMap: resultMap)
           }
           set {
-            resultMap.updateValue(newValue, forKey: "__typename")
-          }
-        }
-
-        public var language: String {
-          get {
-            return resultMap["language"]! as! String
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "language")
-          }
-        }
-
-        public var title: String? {
-          get {
-            return resultMap["title"] as? String
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "title")
-          }
-        }
-
-        public var description: String? {
-          get {
-            return resultMap["description"] as? String
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "description")
-          }
-        }
-
-        public var submitButtonTitle: String? {
-          get {
-            return resultMap["submitButtonTitle"] as? String
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "submitButtonTitle")
-          }
-        }
-
-        public var skipButtonTitle: String? {
-          get {
-            return resultMap["skipButtonTitle"] as? String
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "skipButtonTitle")
-          }
-        }
-
-        public var completionButtonTitle: String? {
-          get {
-            return resultMap["completionButtonTitle"] as? String
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "completionButtonTitle")
-          }
-        }
-
-        public var completionTitle: String? {
-          get {
-            return resultMap["completionTitle"] as? String
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "completionTitle")
-          }
-        }
-
-        public var completionDescription: String? {
-          get {
-            return resultMap["completionDescription"] as? String
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "completionDescription")
-          }
-        }
-      }
-
-      public struct Field: GraphQLSelectionSet {
-        public static let possibleTypes: [String] = ["FormField"]
-
-        public static var selections: [GraphQLSelection] {
-          return [
-            GraphQLField("__typename", type: .nonNull(.scalar(String.self))),
-            GraphQLField("type", type: .scalar(String.self)),
-            GraphQLField("key", type: .scalar(String.self)),
-            GraphQLField("validationRegex", type: .scalar(String.self)),
-            GraphQLField("required", type: .scalar(Bool.self)),
-            GraphQLField("sortId", type: .scalar(Int.self)),
-            GraphQLField("autoFill", type: .object(AutoFill.selections)),
-            GraphQLField("translations", type: .list(.object(Translation.selections))),
-            GraphQLField("selectOptions", type: .list(.object(SelectOption.selections))),
-          ]
-        }
-
-        public private(set) var resultMap: ResultMap
-
-        public init(unsafeResultMap: ResultMap) {
-          self.resultMap = unsafeResultMap
-        }
-
-        public init(type: String? = nil, key: String? = nil, validationRegex: String? = nil, `required`: Bool? = nil, sortId: Int? = nil, autoFill: AutoFill? = nil, translations: [Translation?]? = nil, selectOptions: [SelectOption?]? = nil) {
-          self.init(unsafeResultMap: ["__typename": "FormField", "type": type, "key": key, "validationRegex": validationRegex, "required": `required`, "sortId": sortId, "autoFill": autoFill.flatMap { (value: AutoFill) -> ResultMap in value.resultMap }, "translations": translations.flatMap { (value: [Translation?]) -> [ResultMap?] in value.map { (value: Translation?) -> ResultMap? in value.flatMap { (value: Translation) -> ResultMap in value.resultMap } } }, "selectOptions": selectOptions.flatMap { (value: [SelectOption?]) -> [ResultMap?] in value.map { (value: SelectOption?) -> ResultMap? in value.flatMap { (value: SelectOption) -> ResultMap in value.resultMap } } }])
-        }
-
-        public var __typename: String {
-          get {
-            return resultMap["__typename"]! as! String
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "__typename")
-          }
-        }
-
-        public var type: String? {
-          get {
-            return resultMap["type"] as? String
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "type")
-          }
-        }
-
-        public var key: String? {
-          get {
-            return resultMap["key"] as? String
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "key")
-          }
-        }
-
-        public var validationRegex: String? {
-          get {
-            return resultMap["validationRegex"] as? String
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "validationRegex")
-          }
-        }
-
-        public var `required`: Bool? {
-          get {
-            return resultMap["required"] as? Bool
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "required")
-          }
-        }
-
-        public var sortId: Int? {
-          get {
-            return resultMap["sortId"] as? Int
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "sortId")
-          }
-        }
-
-        public var autoFill: AutoFill? {
-          get {
-            return (resultMap["autoFill"] as? ResultMap).flatMap { AutoFill(unsafeResultMap: $0) }
-          }
-          set {
-            resultMap.updateValue(newValue?.resultMap, forKey: "autoFill")
-          }
-        }
-
-        public var translations: [Translation?]? {
-          get {
-            return (resultMap["translations"] as? [ResultMap?]).flatMap { (value: [ResultMap?]) -> [Translation?] in value.map { (value: ResultMap?) -> Translation? in value.flatMap { (value: ResultMap) -> Translation in Translation(unsafeResultMap: value) } } }
-          }
-          set {
-            resultMap.updateValue(newValue.flatMap { (value: [Translation?]) -> [ResultMap?] in value.map { (value: Translation?) -> ResultMap? in value.flatMap { (value: Translation) -> ResultMap in value.resultMap } } }, forKey: "translations")
-          }
-        }
-
-        public var selectOptions: [SelectOption?]? {
-          get {
-            return (resultMap["selectOptions"] as? [ResultMap?]).flatMap { (value: [ResultMap?]) -> [SelectOption?] in value.map { (value: ResultMap?) -> SelectOption? in value.flatMap { (value: ResultMap) -> SelectOption in SelectOption(unsafeResultMap: value) } } }
-          }
-          set {
-            resultMap.updateValue(newValue.flatMap { (value: [SelectOption?]) -> [ResultMap?] in value.map { (value: SelectOption?) -> ResultMap? in value.flatMap { (value: SelectOption) -> ResultMap in value.resultMap } } }, forKey: "selectOptions")
-          }
-        }
-
-        public struct AutoFill: GraphQLSelectionSet {
-          public static let possibleTypes: [String] = ["AutoFill"]
-
-          public static var selections: [GraphQLSelection] {
-            return [
-              GraphQLField("__typename", type: .nonNull(.scalar(String.self))),
-              GraphQLField("type", type: .scalar(String.self)),
-              GraphQLField("field", type: .scalar(String.self)),
-            ]
-          }
-
-          public private(set) var resultMap: ResultMap
-
-          public init(unsafeResultMap: ResultMap) {
-            self.resultMap = unsafeResultMap
-          }
-
-          public init(type: String? = nil, field: String? = nil) {
-            self.init(unsafeResultMap: ["__typename": "AutoFill", "type": type, "field": field])
-          }
-
-          public var __typename: String {
-            get {
-              return resultMap["__typename"]! as! String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "__typename")
-            }
-          }
-
-          public var type: String? {
-            get {
-              return resultMap["type"] as? String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "type")
-            }
-          }
-
-          public var field: String? {
-            get {
-              return resultMap["field"] as? String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "field")
-            }
-          }
-        }
-
-        public struct Translation: GraphQLSelectionSet {
-          public static let possibleTypes: [String] = ["FormTranslation"]
-
-          public static var selections: [GraphQLSelection] {
-            return [
-              GraphQLField("__typename", type: .nonNull(.scalar(String.self))),
-              GraphQLField("title", type: .scalar(String.self)),
-              GraphQLField("language", type: .nonNull(.scalar(String.self))),
-              GraphQLField("description", type: .scalar(String.self)),
-              GraphQLField("completionDescription", type: .scalar(String.self)),
-              GraphQLField("skipButtonTitle", type: .scalar(String.self)),
-              GraphQLField("completionTitle", type: .scalar(String.self)),
-              GraphQLField("completionButtonTitle", type: .scalar(String.self)),
-              GraphQLField("submitButtonTitle", type: .scalar(String.self)),
-            ]
-          }
-
-          public private(set) var resultMap: ResultMap
-
-          public init(unsafeResultMap: ResultMap) {
-            self.resultMap = unsafeResultMap
-          }
-
-          public init(title: String? = nil, language: String, description: String? = nil, completionDescription: String? = nil, skipButtonTitle: String? = nil, completionTitle: String? = nil, completionButtonTitle: String? = nil, submitButtonTitle: String? = nil) {
-            self.init(unsafeResultMap: ["__typename": "FormTranslation", "title": title, "language": language, "description": description, "completionDescription": completionDescription, "skipButtonTitle": skipButtonTitle, "completionTitle": completionTitle, "completionButtonTitle": completionButtonTitle, "submitButtonTitle": submitButtonTitle])
-          }
-
-          public var __typename: String {
-            get {
-              return resultMap["__typename"]! as! String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "__typename")
-            }
-          }
-
-          public var title: String? {
-            get {
-              return resultMap["title"] as? String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "title")
-            }
-          }
-
-          public var language: String {
-            get {
-              return resultMap["language"]! as! String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "language")
-            }
-          }
-
-          public var description: String? {
-            get {
-              return resultMap["description"] as? String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "description")
-            }
-          }
-
-          public var completionDescription: String? {
-            get {
-              return resultMap["completionDescription"] as? String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "completionDescription")
-            }
-          }
-
-          public var skipButtonTitle: String? {
-            get {
-              return resultMap["skipButtonTitle"] as? String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "skipButtonTitle")
-            }
-          }
-
-          public var completionTitle: String? {
-            get {
-              return resultMap["completionTitle"] as? String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "completionTitle")
-            }
-          }
-
-          public var completionButtonTitle: String? {
-            get {
-              return resultMap["completionButtonTitle"] as? String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "completionButtonTitle")
-            }
-          }
-
-          public var submitButtonTitle: String? {
-            get {
-              return resultMap["submitButtonTitle"] as? String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "submitButtonTitle")
-            }
-          }
-        }
-
-        public struct SelectOption: GraphQLSelectionSet {
-          public static let possibleTypes: [String] = ["FormSelectOption"]
-
-          public static var selections: [GraphQLSelection] {
-            return [
-              GraphQLField("__typename", type: .nonNull(.scalar(String.self))),
-              GraphQLField("title", type: .scalar(String.self)),
-              GraphQLField("value", type: .scalar(String.self)),
-              GraphQLField("iconUrl", type: .scalar(String.self)),
-            ]
-          }
-
-          public private(set) var resultMap: ResultMap
-
-          public init(unsafeResultMap: ResultMap) {
-            self.resultMap = unsafeResultMap
-          }
-
-          public init(title: String? = nil, value: String? = nil, iconUrl: String? = nil) {
-            self.init(unsafeResultMap: ["__typename": "FormSelectOption", "title": title, "value": value, "iconUrl": iconUrl])
-          }
-
-          public var __typename: String {
-            get {
-              return resultMap["__typename"]! as! String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "__typename")
-            }
-          }
-
-          public var title: String? {
-            get {
-              return resultMap["title"] as? String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "title")
-            }
-          }
-
-          public var value: String? {
-            get {
-              return resultMap["value"] as? String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "value")
-            }
-          }
-
-          public var iconUrl: String? {
-            get {
-              return resultMap["iconUrl"] as? String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "iconUrl")
-            }
+            resultMap += newValue.resultMap
           }
         }
       }
@@ -1004,25 +404,7 @@ public extension ApolloType {
       public static var selections: [GraphQLSelection] {
         return [
           GraphQLField("__typename", type: .nonNull(.scalar(String.self))),
-          GraphQLField("id", type: .nonNull(.scalar(GraphQLID.self))),
-          GraphQLField("label", type: .scalar(String.self)),
-          GraphQLField("name", type: .scalar(String.self)),
-          GraphQLField("status", type: .scalar(String.self)),
-          GraphQLField("isDefault", type: .scalar(Bool.self)),
-          GraphQLField("description", type: .scalar(String.self)),
-          GraphQLField("imageUrl", type: .scalar(String.self)),
-          GraphQLField("mapImageUrl", type: .scalar(String.self)),
-          GraphQLField("geoLat", type: .scalar(Double.self)),
-          GraphQLField("geoLong", type: .scalar(Double.self)),
-          GraphQLField("geoLatNorthWest", type: .scalar(Double.self)),
-          GraphQLField("geoLongNorthWest", type: .scalar(Double.self)),
-          GraphQLField("geoLatSouthEast", type: .scalar(Double.self)),
-          GraphQLField("geoLongSouthEast", type: .scalar(Double.self)),
-          GraphQLField("city", type: .scalar(String.self)),
-          GraphQLField("externalId", type: .scalar(String.self)),
-          GraphQLField("createdAt", type: .scalar(String.self)),
-          GraphQLField("updatedAt", type: .scalar(String.self)),
-          GraphQLField("venueIconUrl", type: .scalar(String.self)),
+          GraphQLFragmentSpread(FragmentVenue.self),
         ]
       }
 
@@ -1045,174 +427,29 @@ public extension ApolloType {
         }
       }
 
-      public var id: GraphQLID {
+      public var fragments: Fragments {
         get {
-          return resultMap["id"]! as! GraphQLID
+          return Fragments(unsafeResultMap: resultMap)
         }
         set {
-          resultMap.updateValue(newValue, forKey: "id")
+          resultMap += newValue.resultMap
         }
       }
 
-      public var label: String? {
-        get {
-          return resultMap["label"] as? String
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "label")
-        }
-      }
+      public struct Fragments {
+        public private(set) var resultMap: ResultMap
 
-      public var name: String? {
-        get {
-          return resultMap["name"] as? String
+        public init(unsafeResultMap: ResultMap) {
+          self.resultMap = unsafeResultMap
         }
-        set {
-          resultMap.updateValue(newValue, forKey: "name")
-        }
-      }
 
-      public var status: String? {
-        get {
-          return resultMap["status"] as? String
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "status")
-        }
-      }
-
-      public var isDefault: Bool? {
-        get {
-          return resultMap["isDefault"] as? Bool
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "isDefault")
-        }
-      }
-
-      public var description: String? {
-        get {
-          return resultMap["description"] as? String
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "description")
-        }
-      }
-
-      public var imageUrl: String? {
-        get {
-          return resultMap["imageUrl"] as? String
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "imageUrl")
-        }
-      }
-
-      public var mapImageUrl: String? {
-        get {
-          return resultMap["mapImageUrl"] as? String
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "mapImageUrl")
-        }
-      }
-
-      public var geoLat: Double? {
-        get {
-          return resultMap["geoLat"] as? Double
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "geoLat")
-        }
-      }
-
-      public var geoLong: Double? {
-        get {
-          return resultMap["geoLong"] as? Double
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "geoLong")
-        }
-      }
-
-      public var geoLatNorthWest: Double? {
-        get {
-          return resultMap["geoLatNorthWest"] as? Double
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "geoLatNorthWest")
-        }
-      }
-
-      public var geoLongNorthWest: Double? {
-        get {
-          return resultMap["geoLongNorthWest"] as? Double
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "geoLongNorthWest")
-        }
-      }
-
-      public var geoLatSouthEast: Double? {
-        get {
-          return resultMap["geoLatSouthEast"] as? Double
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "geoLatSouthEast")
-        }
-      }
-
-      public var geoLongSouthEast: Double? {
-        get {
-          return resultMap["geoLongSouthEast"] as? Double
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "geoLongSouthEast")
-        }
-      }
-
-      public var city: String? {
-        get {
-          return resultMap["city"] as? String
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "city")
-        }
-      }
-
-      public var externalId: String? {
-        get {
-          return resultMap["externalId"] as? String
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "externalId")
-        }
-      }
-
-      public var createdAt: String? {
-        get {
-          return resultMap["createdAt"] as? String
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "createdAt")
-        }
-      }
-
-      public var updatedAt: String? {
-        get {
-          return resultMap["updatedAt"] as? String
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "updatedAt")
-        }
-      }
-
-      public var venueIconUrl: String? {
-        get {
-          return resultMap["venueIconUrl"] as? String
-        }
-        set {
-          resultMap.updateValue(newValue, forKey: "venueIconUrl")
+        public var fragmentVenue: FragmentVenue {
+          get {
+            return FragmentVenue(unsafeResultMap: resultMap)
+          }
+          set {
+            resultMap += newValue.resultMap
+          }
         }
       }
     }
