+++ Core/GraphQL/API/sell_getFulfilmentPointCategories.graphql.swift
@@ -20,8 +20,6 @@ public extension ApolloType {
               ...FragmentFulfilmentPointCategory
             }
             nextPage
-            lastPage
-            firstPage
           }
         }
       }
@@ -121,8 +119,6 @@ public extension ApolloType {
               GraphQLField("__typename", type: .nonNull(.scalar(String.self))),
               GraphQLField("edges", type: .list(.object(Edge.selections))),
               GraphQLField("nextPage", type: .scalar(Int.self)),
-              GraphQLField("lastPage", type: .scalar(Int.self)),
-              GraphQLField("firstPage", type: .scalar(Int.self)),
             ]
           }
 
@@ -132,8 +128,8 @@ public extension ApolloType {
             self.resultMap = unsafeResultMap
           }
 
-          public init(edges: [Edge?]? = nil, nextPage: Int? = nil, lastPage: Int? = nil, firstPage: Int? = nil) {
-            self.init(unsafeResultMap: ["__typename": "FulfilmentPointCategoryEdge", "edges": edges.flatMap { (value: [Edge?]) -> [ResultMap?] in value.map { (value: Edge?) -> ResultMap? in value.flatMap { (value: Edge) -> ResultMap in value.resultMap } } }, "nextPage": nextPage, "lastPage": lastPage, "firstPage": firstPage])
+          public init(edges: [Edge?]? = nil, nextPage: Int? = nil) {
+            self.init(unsafeResultMap: ["__typename": "FulfilmentPointCategoryEdge", "edges": edges.flatMap { (value: [Edge?]) -> [ResultMap?] in value.map { (value: Edge?) -> ResultMap? in value.flatMap { (value: Edge) -> ResultMap in value.resultMap } } }, "nextPage": nextPage])
           }
 
           public var __typename: String {
@@ -160,24 +156,6 @@ public extension ApolloType {
             }
             set {
               resultMap.updateValue(newValue, forKey: "nextPage")
-            }
-          }
-
-          public var lastPage: Int? {
-            get {
-              return resultMap["lastPage"] as? Int
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "lastPage")
-            }
-          }
-
-          public var firstPage: Int? {
-            get {
-              return resultMap["firstPage"] as? Int
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "firstPage")
             }
           }
