+++ Core/GraphQL/GraphQLFiles/sell_fragment_fulfilmentPoint.graphql
@@ -15,75 +15,8 @@ fragment FragmentFulfilmentPoint on FulfilmentPoint {
     language
     title
   }
-  seatForm {
-    id
-    reference
-    imageUrl
-    completionButtonUrl
-    completionButtonTitle
-    showCompletionDate
-    allowUpdate
-    refreshOnSuccess
-    validationIntegration
-    expiresIn
-    translations {
-      language
-      title
-      description
-      submitButtonTitle
-      skipButtonTitle
-      completionButtonTitle
-      completionTitle
-      completionDescription
-    }
-    fields {
-      type
-      key
-      validationRegex
-      required
-      sortId
-      autoFill {
-        type
-        field
-      }
-      translations {
-        title
-        language
-        description
-        completionDescription
-        skipButtonTitle
-        completionTitle
-        completionButtonTitle
-        submitButtonTitle
-      }
-      selectOptions {
-        title
-        value
-        iconUrl
-      }
-    }
-  }
+  seatForm { ...FragmentForm }
   categories { ...FragmentFulfilmentPointCategory }
-  venue {
-    id
-    label
-    name
-    status
-    isDefault
-    description
-    imageUrl
-    mapImageUrl
-    geoLat
-    geoLong
-    geoLatNorthWest
-    geoLongNorthWest
-    geoLatSouthEast
-    geoLongSouthEast
-    city
-    externalId
-    createdAt
-    updatedAt
-    venueIconUrl
-  }
+  venue { ...FragmentVenue }
   timeslots { ...FragmentTimeslot }
 }
