+++ Core/GraphQL/GraphQLFiles/sell_fragment_orders.graphql
@@ -3,6 +3,4 @@ fragment FragmentOrders on OrderEdge {
     ...FragmentOrder
   }
   nextPage
-  lastPage
-  firstPage
 }
