+++ Core/GraphQL/GraphQLFiles/sell_getFulfilmentPointCategories.graphql
@@ -3,8 +3,6 @@ query getFulfilmentPointCategories($pageSize: Int!, $page: Int = 1) {
     fulfilmentPointCategories(pageSize: $pageSize, page: $page) {
       edges { ...FragmentFulfilmentPointCategory }
       nextPage
-      lastPage
-      firstPage
     }
   }
 }
