+++ Core/GraphQL/GraphQLFiles/sell_getFulfilmentPoints.graphql
@@ -1,7 +1,6 @@
 query getFulfilmentPoints($pageSize: Int!, $page: Int = 1, $filters: FulfilmentPointFilter = null) {
   getFulfilmentPoints(page: $page, pageSize: $pageSize, filters: $filters) {
-    edges {
-      ...FragmentFulfilmentPoint
-  	}
+    edges { ...FragmentFulfilmentPoint }
+    nextPage
   }
 }
