+++ Core/GraphQL/GraphQLFiles/sell_getProducts.graphql
@@ -1,4 +1,4 @@
-query GetProducts($pageSize: Int!, $page: Int, $filters: ProductFilter) {
+query getProducts($pageSize: Int!, $page: Int, $filters: ProductFilter) {
   getProducts(pageSize: $pageSize, page: $page, filters: $filters) {
     ...FragmentProducts
   }
